Support usage outside of docker-compose.
The primary change is to move the mapping of container names out of the
core functions and only do it during top-level configuration loading and
configuration mangling. The parameterize the compose specific loading
to only happen if the new --compose-file flag is given.

Add third example that demonstrates launching a standalone conlink
network container that connects to a pre-existing docker container.