Allow network-settings key in schema top-level.
The 'network-settings' key is not directly used by conlink, but the
intention is that with a larger configuration it is useful to have an
arbitrary blob that captures other network information in the same place
as the rest of the network config. For example, this could be shell
variables that are extracted and used by other scripts.