mininet: use regular Host instead of CPULimitedHost
This could probably be parameterized, but for now there is no to specify
the CPU limits for a host via the declarative format and the
CPULimitedHost class requires the ability to create cgroups which limits
where it can be used (e.g. enables rootless podman on a cgroup v1 system).