Do not start ovs-vswitchd with mlockall.
The mlockall parameter keeps the vswitchd process from being swapped out
in a low memory condition. However, the parameter also requires elevated
permissions to invoke and this causes vswitch to crash in a context
without those permissions (such as rootless podman). For now, prefer
the ability to run in more contexts than the possible future issue of
swapping in low memory.