Add test6: like test5 with CloudFormation automation
Add `examples/test6-cfn.yaml`, a cloudformation template that deploys
two instances that boot up and install all the dependencies for running
conlink and then launch conlink using `examples/test5-geneve.yaml`
network configuration to create a connection between the two instances.

The cloudformation template provisions two instances into an existing
VPC/subnet. The userdata of the instances performs the following:
- install podman and docker apt repos
- install podman and docker
- load openvswitch and geneve kernel modules
- enable dbus/loginctl linger for the ubuntu user so that rootless
  podman can use systemd for cgroup management.
- pull down the conlink podman image and clone the conlink repo
- signal the cloudformation template that instance has been provisioned.
- start conlink as the ubuntu user using the
  `examples/test5-geneve.yaml` network config and pointing at the IP of
  the ENI of the opposite instance for the value of REMOTE.

The provisioning of an instance only takes about 1 minute. However, the
rest of the userdata up to the signalling to CFN takes another 7 minutes
or so. The 7 minutes of configuration could be eliminated by creating an
AMI with all the steps prior to the signalling already installed and
using that instead.