Add env var interpolation in inline network config.
Previously, variables in the main compose file configuration would be
interpolation by docker-compose but those same variables in the inline
section would not be interpolation by conlink. This now allows
interpolation (in docker-compose style with default values and require
values syntax) in the whole compose file including the inline conlink
configuration.

Update test1 to include environment variables for the last octet of each
host's IP address (with a default of "100" for all three of them).