Support multiple compose and network files.
This adds support for merging network configuration from multiple
locations. Both --compose-file and --network-file can be specified
multiple times and each can also accept a colon ':' separate list of
files. In addition, the 'x-network' property can now be specified in
multiple places within a compose files; at the top-level and within one
or more services (and not just the conlink service itself).

The directory `examples/test7-multiple/` that demonstrate how a running
configuration can be constructed from multiple files and can be defined
modularly so that smaller configurations can be run using a subset of
the files.