Ability to move/adopt host interfaces into containers.
This adds a top-level 'interfaces' key that is a list of host interfaces
that should be moved into containers. Interface movement/adoption
happens after veth links but before commands are run for the target
container.

Each item in the interfaces list has 3 required propeties:
    - container: container specification (same as links and commands)
    - host-intf: name of the host interface to move/adopt
    - intf:      rename interface to this in the container

There are also 3 optional properties:
    - ipvlan: if true, then a ipvlan sub-interface will be created and moved instead
    - ip:     IP/CIDR to assign to the interface after it is moved
    - nat:    add static NAT (DNAT+SNAT) from the interface to this address

To implement this `move-intf.sh` is a new shell script similar to
`veth-link.sh` that is a fairly general interface movement command. The
args/opts are map mostly 1-to-1 with the interfaces specification
however this command also takes a source PID that allows interfaces to
be moved from any netns to any other netns.