Use container ID when outer interface too long
Try to use CONTAINER-INTF or SERVICE_INDEX-INTF name for outer
interface. If the length of the name is over 15 then fallback to
cID-INTF (with ID and INTF truncated so the total name is never more
than 15).

This should give unique and human readable name for most cases and at
minimum unique names for the case where the total length exceeds the
interface name limit (15 characters).

This also means generating the outer interface name can only
be done when we get a link event and we can't pregenerate the outer
interface names at configuration load time.