Suport live scale change. Iterate IP/MAC addresses.
The ip and mac address link settings become the starting point for
compose services. If there is only a single instance/replica of
a service (scale: 1), then there is no behavior change. However, for
container indexes greater than 1, the IP and MAC address will be
iterated from the start. So for a service where the IP is defined in
a link as "192.168.1.1/16", replica five (index=5) would get an actual
IP settings of "192.168.1.5/16".

Also, remove the link from the domain/switch rather than just deleting
the link when a container goes away. Otherwise in OVS bridge-mode, the
bridge still has tracking for the link even though the physical
interface is deleted. This allows scale down (and back up).

Example of using scale:

    docker-compose -f examples/test2-compose.yaml up -d
    docker-compose -f examples/test2-compose.yaml up -d --scale node=3
    docker-compose -f examples/test2-compose.yaml up -d --scale node=2
    docker-compose -f examples/test2-compose.yaml up -d --scale node=3