Drop sub-commands. Move loggers/bridge-mode to ctx
Move away from the sub-commands entirely and make 'inner' the only
command/mode.

Restructure things to put the logger functions in the ctx state var. In
addition, copy bridge-mode into the ctx to simplify the various shell
command wrapper functions. At some point we might want the bridge-mode
to be part of the domain definition so that the same conlink instances
can support multiple bridge modes simultaneously (with a default for
links that don't specify).