veth-link.sh: fix interface name collision
The previous ip link command was this:
  ip link add ${IF0} netns ns${PID0} type veth peer ${IF1} netns ns${PID1}

It's not clear from the documentation, but this is not an idempotent
command. The interfaces endpoints are created in the current name space
and then moved to the destination namespaces as separate steps. This
means that the IF0 and IF1 interface names can collide with interface
names in the parent namespace even if they wouldn't collide with an
interface in the target namespace.

To fix this issue, use temporary interface names with a random suffix in
the parent namespace and then rename them after moving them to the
target namespaces.