Add --compose-project. Live project match in handle-event
Instead of filtering and listing only those containers that match the
context of the container during startup, check the compose properties of
the event container at runtime against the compose definitions. To
This allows an external/separate instance of conlink to match against
the services of a docker-compose instances. To support this, the
--compose-project is added. This basically specifies how to resolve the
link service when conlink is not running as one of the services of the
compose instance (which is where the project name and project
working_dir is queried if --compose-project is not specified).