conlink-start.sh: disabled host bridge net for conlink container
For some unknown reason, when the conlink container is first started,
there seems to be interference with subsequent ip link actions that
causes "RTNETLINK answers: File exists" errors for about half the `ip
link add` conlink commands. Once the container has been running for
a while (i.e. bringing up and down the connected containers) the errors
go away so it seems to indicate that docker is doing something early in
container startup of the conlink container that is causing these issues.
So just disable host bridge networking for the conlink container. This
mostly shouldn't have an impact on usage of conlink since external
connectivity from the overlay network would usually be via a separate
drain style container and not via conlink (where the interfaces
endpoints don't generally have an IP anyways to be able to route traffic
into the connected containers).