Add commands. Default service key. Restructure tests.
Add support for running container commands after all the links are
created for that container.

When the x-network link properties are defined under a docker-compose
service, then use the service name/key as the default value of :service
for all links defined under that service. The value is a default and can
be overridden (there may be situations where it makes sense to group
links under one service that apply to a different service).

Combine examples/test4* into the test3 section of the README since it's
the same test but with podman instead.

Rename the multiple compose file example test7 to test4. This is more
basic functionality than test5, etc so move it earlier. This example
utilizes the commands functionality to start a web server that is
queried in the README walkthrough.