Add link :creating and :deleting status values.
When a link is starting to be created, set it's status state to
:creating and then move it to :created once it is fully created.
Likewise, when a link is starting to be deleted, set it to :deleting and
to nil when it's fully removed. This shouldn't affect any current status
logic but it does allow future changes to be able to detect links that
are changing but haven't fully resolved yet.