Add check that prints when all links are created.
Checks that the number of live links is equal to the number of service
and container links that are defined and print out "All links
connected". Also, if verbose is set then print out the final
network-state value.

Caveat: For a situation where use docker-compose scale/replicas, this
check will likely fire early.