Move lookups from state to config. Add docstrings.
The containers and services lookup maps aren't actually runtime state so
move them to the network-config data and add the network-config data to
the context (ctx). The network-state is reduced to just the :status
field for the runtime :bridges and :links.

Also add full docstrings to all the functions and clarify the async
startup paragraph of the README.