Rename docker eth0 interface in conlink container.
Reverts a2fda8b6c807b50121119facce4c442db7ac7979 since conlink container
docker networking should now work.

Rename the docker eth0 interface (if it exists) to prevent 'RTNETLINK
answers: File exists' errors during creation of links that use an 'eth0'
device name. This is necessary because even if the netns is specified
with the same link create command, the creation and move does not appear
to be idempotent and results in the conflict.

Update to nbb 1.2.179 to fix an issue with `P/loop` and then use that to
create a `run*` command that runs multiple commands and returns the
final status (or the status of the first erroring command).

Use the `run*` to implement the multi-step process of renaming a network
interface:
    - save routes
    - set interface down
    - rename interface
    - set interface up
    - restore routes