Add tc qdisc NetEm property. README, schema cleanup
The new `netem` property can be used to set NetEm traffic control on an
interface/link (delay, packet loss, duplication, reordering, etc).

See `man 8 netem` for the OPTIONS that can be used in this field.

Also re-arrange the rows/cols of the link table in the README. Add
property tables for tunnels and commands too.

Rename test8 to test7 and define it to be a MTU, MAC, and NetEm
example.