copy.sh should use temp file instead of sed in place
The default behavior of `sed -i` is to create a temp file in the same
directory and then move it into place. If the target file is a mount
point then this will fail. The GNU sed command does suport a `-c` option
to copy instead of move. However, this is not support by busybox sed so
implement the copy behavior ourselves.

Co-authored-by: Joel Martin <joel.martin@viasat.com>