New default "auto" bridge-mode. Falls back to linux bridging.
Add support for "auto" --bridge-mode. This will check if the kernel has
the "openvswitch" kernel module and if not it will automatically fall
back to using standard linux bridging. If "ovs" bridge mode is specified
then conlink will error and stop during startup if the "openvswitch"
module is not loaded.

Add a bridging subsection to "Usage Notes' section of the README that
describes this functionality/behavior.

Refactor the `startup-checks` function to take the full opts and return
an updated opts with the docker, podman, and bridge-mode keys
set/updated based on checks.