Add patch bridge mode (using tc filter action).
Add a new "patch" mode (e.g. bump-in-the-wire) that uses the kernel
module "act_mirred" and a tc ingress filter action to redirect traffic
from one interface to another interface and vice versa. One key
limitation of this "bridge" action is that only two links can be
connected. So while the this mode supports restarts/reconnecting of
containers it does not support arbitrary scalability like the ovs and
linux bridge modes.

Add a mirred node to the test9-compose.yaml example and document how to
use/test it in the README.