[run-tests] refactor test functions for more flexibility.
The function that runs a docker-compose commmand becomes dc_run and
that's all it does (not wait loop). The main test running function that
includes the retry functionality is split out into a non-compose
specific function called do_test. If the WITH_DC environment variable is
set then it uses dc_run to run a command, otherwise it invokes the
arguments in a regular shell call.

The dc_wait and dc_test function become trivial wrappers around do_test.