Support string or array for forward/route attributes.
Previously we only supported a single route and the forward definition
was required to be an array. Add support for specifying multiple
routes (as an array of strings) and make forward work the same way so
that a single forward definition can be specified as a plain string.

This requires modifying the link-add.sh script to support the --route
option being specified multiple times on the command line. The schema is
adjusted to allow string or array of strings in the forward and route
attributes.

Update the test10 example to add a router container and the node
containers to use multiple routes. Add routing tests to the test10
section in run-tests.sh.