Automatically add reverse route for forwards.
Read the docker eth0 IPv4 address and add a reverse (*/32) route to that
address from each container with a port forward destination. Previously,
the forward link had to be the default route in the container or the
user had to use static docker bridge address allocation and then
explicitly add a return route to the container (not typically a workable
solution). So this removes the need for that and should setup a narrow
/32 reverse route just for the docker eth0 address that will be the
source of the forwarded/masqueraded traffic.

This adds a new network function called intf-ipv4-addresses to read the
IPv4 network addresses from an interface.