Fix forwards/routing in linux bridge mode.
For some reason, when using regular Linux bridges, the forward rules
match normal traffic and cause it to be NAT'd/MASQUERADED which breaks
routing of normal traffic (this doesn't happen in ovs mode).

Tighten up the iptables rules so that the forwarded traffic is marked in
a PREROUTING rule and then only the marked traffic is MASQUERADED in
POSTROUTING rule.