mdc support for multiple modes directories
resolve-deps supports multiple (comma-separated) directories for nodes
and can show the resolution in "node=/some/path" format if provided the
`--format=paths` flag option.  Use this to support multiple mdc modes
directories.

Reorder the echo/vecho lines a bit to reduce the processing required to
display the "MODES: .." line to the user.