Stabalize MACs to fix ARP timeouts on restarts
When a container is restarted and links are recreated (without user
specified MAC addresses), other containers that are connected to the
same broadcast domain (bridge, patch) will still have the original MAC
address in their ARP tables which will cause connectivity failures until
the ARP table entries expire and are recreated with the new MAC address.
To avoid this, pre-assign a random MAC address to links when the user
has not specified a MAC address. This MAC address will persist when the
same link is recreated.

Additionally, use "c2" as the first octet of the random MAC addresses.
This will help identify conlink assigned random MAC addresses.