Fix container delete race by saving container state.
When a a container shutdown event ('die') is received, we previously
tried to inspect/query the container state in order to have the
information needed to delete container links and bridge state. However,
occasionally, the container was already completely gone by the time we
try and inspect its state and so we get a 404 error from the inspect
call. To fix this race, we save the container state after we query the
container during the "start" event. When we get a "die" event, then use
this saved context/state instead of inspecting/querying the container
during shutdown/delete.