Store container id for use in link/oif naming fallback
'link-outer-dev normally creates a friendly name for the outer link
device using a combination of service and device name; however, if that
string is look long, the fallback strategy is to use a portion of the
container id. Unfortunately the id key was removed during a refactoring,
which introduced the regression.