Do not wait/block for docker-listen (.getEvents)
In some contexts (certain docker engine versions?), the dockerode client
call to .getEvents blocks (until the first event) instead of immediately
returning an event-stream object. We don't actually need this
event-stream object so just wrap the call to docker-listen in a vector
so that promesa won't await/block on it.