mdc: add MDC_MODE_* variables to env file.
These are similar to MODE_* profile variables but are set in the
environment instead. The prefix MDC_ is added just to reduce the chance
of colliding in a broader environment context.

Important Note: for the mode environment variables, all non-alphanumeric
characters in the mode name are changed to lowercase for the variable
name.