Add static Rust wait/copy/echo commands.
There are scenarios where the service images do not have the facilities
to be able to run the wait.sh and copy.sh commands. For example, "FROM
scratch" docker images can have a single static executable that defines
the service. In this case, the image probably does not even have a shell
command.

Adds wait and copy commands (and a simple echo command) that are built
into statically linked and standalone executables. These can then be
volumed mounted into any docker image and used in the command line
without needing to install any other dependencies in the image. This allows
the original upstream images to be used without need to extending them with
a Dockerfile which avoids many versioning problems and the extra build
process/time needed to add dependencies into an image.

The commands are built and installed into `/utils` within the conlink
image. These can be copied out and used by other services. Refer to the
`build` service in the `test/utils.yaml` compose file for an example of this.

The `test/utils.yaml` file is a docker compose file that will build and
test the static executable. It can be used like this:
```
docker compose -f test/utils.yaml build
docker compose -f test/utils.yaml up --force-recreate --build
```

All the services will run and exit with success and the overall compose
up command will exit with success (within about 10 seconds). If
something does not work then the compose up process will either hang or
will exit with a failure code.