rust/copy: exec cmd rather than just calling it.
The wait command was already doing this but copy was not. For many
services this is not an issue. However, systemd init processes expect to
be the first process (pid 1), and fail with this error if they are not:
    Couldn't find an alternative telinit implementation to spawn.

Also, add running command print to wait to be consistent with copy (and
with the original wait.sh).