fix chown syntax
Some `chown` binaries seem to support the `.` notation, but a few don't.  I haven't looked into why that is.  Since `:` is the arguably ubiquitous, conventional, and universal (?) syntax, I'm changing it to head in that direction.  That said, usage of `copy.sh` should probably be phased out, anyway, and the from-rust-built binary `copy` should be used instead.